画像アップロード時にログインユーザーのidを取得しfileテーブルのuser_idカラムに格納 削除の際はログインユーザのidを取得し、選択画像のuser_idと一致しなければdelete不能、エラーメッセージを表示する

+++ laravel/app/Http/Controllers/FileController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\File;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Auth;
 
 
 class FileController extends Controller
@@ -24,6 +25,7 @@ class FileController extends Controller
             // ファイルパスをモデルに保存
             $file = new File;
             $file->file_path = str_replace('public/image/', '', $filePath);
+            $file->user_id = Auth::id();
             $file->save();
 
             // 成功した場合のリダイレクト
@@ -38,13 +40,24 @@ class FileController extends Controller
     {
         // データベースからすべてのファイルを取得してビューに渡す
         $files = File::all();
+
         return view('user.file', compact('files'));
     }
 
 
     public function delete($id)
     {
+
+
         $file = File::find($id);
+
+        $userId = Auth::id();
+        // ログインユーザーが画像の所有者かどうかを確認します
+        if ($file->user_id !== $userId) {
+            return redirect()->back()->with('error', 'この画像は削除できません。');
+
+        }
+
         Storage::delete('storage/image/'. $file->file_path);
 
         $file->delete();
